+++ tools/CMakeLists.txt
@@ -36,4 +36,4 @@ target_link_libraries(yul-phaser PRIVATE phaser)
 
 install(TARGETS yul-phaser DESTINATION "${CMAKE_INSTALL_BINDIR}")
 
-add_subdirectory(yulASTComparator)
+add_subdirectory(yuldiff)
